Move lib to the same dir. Update rulesDirectory
