+++ packages/tslint-plugin-wix-style-react/src/utils.js
